implement proper slug linking
