+++ src/app/form/[slug]/attachments/page.tsx
@@ -1,19 +1,44 @@
+// "use client";
+// import * as React from 'react';
+// import AttachmentsList from "./AttachmentsList";
+// import { useRouter, usePathname } from "next/navigation";
+// import Link from 'next/link';
+
+// const Attachments = () => {
+//   const router = useRouter()
+//   const pathname = usePathname();
+//   console.log(pathname)
+
+//   return (
+//     <div className="govuk-width-container">
+//       <div className="govuk-grid-row">
+//         <div className="govuk-grid-column-two-thirds">
+//           <h1 className="govuk-heading-xl" style={{marginBottom: '0'}}>(Form Name)</h1>
+//           <span className="govuk-caption-m">Attachments</span>
+//           <div style={{ width: "100%", height: ".25rem", backgroundColor: "#AA8CAE", marginTop: "25px", marginBottom: "15px" }}></div>
+
+//           <button className="" onClick={() => router.push("/receipt")}>
+//             Upload
+//           </button>
+//         </div>
+//       </div>
+//     </div>
+//   );
+// }
+
+// export default Attachments; 
+
 "use client";
 import * as React from 'react';
 import AttachmentsList from "./AttachmentsList";
+import { useRouter, usePathname, useParams } from "next/navigation";
+import Link from 'next/link';
 
 const Attachments = () => {
-  const [files, setFiles] = React.useState<File[]>([]);
-
-  const handleFileUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (!event.target.files) return;
-    const newFiles = Array.from(event.target.files);
-    setFiles((prevFiles) => [...prevFiles, ...newFiles]);
-  };
-
-  const deleteFile = (fileName: string) => {
-    setFiles((prevFiles) => prevFiles.filter((file) => file.name !== fileName));
-  };
+  const router = useRouter();
+  const params = useParams();
+  const { slug } = params; // assuming the page route is form/[slug]/attachments
+  console.log("Current slug:", slug);
 
   return (
     <div className="govuk-width-container">
@@ -21,15 +46,19 @@ const Attachments = () => {
         <div className="govuk-grid-column-two-thirds">
           <h1 className="govuk-heading-xl" style={{marginBottom: '0'}}>(Form Name)</h1>
           <span className="govuk-caption-m">Attachments</span>
-          <div style={{ width: "100%", height: ".25rem", backgroundColor: "#AA8CAE", marginTop: "25px", marginBottom: "15px" }}></div>
+          <div
+            style={{
+              width: "100%",
+              height: ".25rem",
+              backgroundColor: "#AA8CAE",
+              marginTop: "25px",
+              marginBottom: "15px",
+            }}
+          ></div>
 
-          <AttachmentsList files={files} onDelete={deleteFile} onUpload={handleFileUpload}/>
-
-          <button className="govuk-button" data-module="govuk-button">
-            Save
-          </button>
-          <button className="govuk-button govuk-button--warning" data-module="govuk-button">
-            Cancel
+          {/* Now push to the receipt page using the current slug */}
+          <button onClick={() => router.push(`/form/${slug}/receipt`)}>
+            Upload
           </button>
         </div>
       </div>
@@ -37,4 +66,4 @@ const Attachments = () => {
   );
 }
 
-export default Attachments; 
+export default Attachments;
