+++ src/app/form/[slug]/receipt/page.tsx
@@ -1,0 +1,5 @@
+import ReceiptPage from "@/app/components/receipts/receipt-page";
+
+export default function Receipt() {
+  return <ReceiptPage />;
+}
