--- src/app/receipt/page.tsx
@@ -1,5 +1,0 @@
-import ReceiptPage from "@/app/components/receipts/receiptPage";
-
-export default function Receipt() {
-  return <ReceiptPage />;
-}
